extract functions to file_helper
